(Chore) Add Netlify build
