+++ internals/webpack/webpack.dev.babel.js
@@ -1,18 +1,9 @@
-const fs = require('fs');
 const path = require('path');
 const webpack = require('webpack');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
 const CircularDependencyPlugin = require('circular-dependency-plugin');
 const webpackBase = require('./webpack.base.babel');
-const config = require('../../environment.conf.json');
-
-const configPlaceholder = '$SIGNALS_CONFIG';
-const configString = JSON.stringify(config);
-const indexFile = path.join(__dirname, '..', '..', 'src', 'index.html');
-const templateContent = fs
-  .readFileSync(indexFile)
-  .toString()
-  .replace(configPlaceholder, configString);
+const template = require('./template');
 
 module.exports = webpackBase({
   mode: 'development',
@@ -40,7 +31,7 @@ module.exports = webpackBase({
     new webpack.HotModuleReplacementPlugin(), // Tell webpack we want hot reloading
     new HtmlWebpackPlugin({
       inject: true, // Inject all files that are generated by webpack, e.g. bundle.js
-      templateContent,
+      ...template,
     }),
     new CircularDependencyPlugin({
       exclude: /a\.js|node_modules/, // exclude node_modules
